Add files via upload
Report :
Started at - 8:30 PM
End time -9:20 PM

Doubt :
We gave the datatype for the two number (on which the operation is to be performed) as integer, since you mentioned it to be the same. But the float type result can not be printed. I hope its okay in case of division operation.